# Conflicts:
#	src/main/java/ch/uzh/ifi/hase/soprafs23/controller/UserController.java
#	src/main/java/ch/uzh/ifi/hase/soprafs23/entity/Food.java
#	src/main/java/ch/uzh/ifi/hase/soprafs23/rest/dto/FoodGetDTO.java
#	src/main/java/ch/uzh/ifi/hase/soprafs23/rest/mapper/DTOMapper.java

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/model/PlayerGuesses.java
@@ -11,5 +11,5 @@ import java.util.Map;
 @Getter
 @Setter
 public class PlayerGuesses {
-    private Map<String,Integer> content;
+    private Map<String,Double> content;
 }
